--- src/test/java/kitchenpos/fixture/MenuGroupFixture.java
@@ -1,9 +1,0 @@
-package kitchenpos.fixture;
-
-import kitchenpos.menugroup.domain.MenuGroup;
-
-public class MenuGroupFixture {
-    public static MenuGroup 추천메뉴 = new MenuGroup(1L, "추천 메뉴");
-    public static MenuGroup 신메뉴 = new MenuGroup(2L, "신메뉴");
-    public static MenuGroup 할인메뉴 = new MenuGroup(3L, "할인 메뉴");
-}
