--- src/test/java/kitchenpos/fixture/MenuProductFixture.java
@@ -1,12 +1,0 @@
-package kitchenpos.fixture;
-
-import kitchenpos.menu.domain.MenuProduct;
-
-import static kitchenpos.fixture.ProductFixture.*;
-
-public class MenuProductFixture {
-    public static MenuProduct 후라이드치킨_한마리_메뉴상품 = new MenuProduct(후라이드치킨.getId(), 1);
-    public static MenuProduct 양념치킨_한마리_메뉴상품 = new MenuProduct(양념치킨.getId(), 1);
-    public static MenuProduct 간장치킨_두마리_메뉴상품 = new MenuProduct(간장치킨.getId(), 2);
-    public static MenuProduct 강정치킨_한마리_메뉴상품 = new MenuProduct(강정치킨.getId(), 1);
-}
