--- src/test/java/kitchenpos/fixture/ProductFixture.java
@@ -1,10 +1,0 @@
-package kitchenpos.fixture;
-
-import kitchenpos.product.domain.Product;
-
-public class ProductFixture {
-    public static Product 후라이드치킨 = new Product(1L, "후라이드치킨", 17000);
-    public static Product 양념치킨 = new Product(2L, "양념치킨", 17000);
-    public static Product 간장치킨 = new Product(3L, "간장치킨", 17000);
-    public static Product 강정치킨 = new Product(4L, "강정치킨", 16000);
-}
